Revert "Styles: Polish MainPage components Layout, update Feed/ folder structure "
