+++ src/components/SignIn/index.jsx
@@ -5,9 +5,12 @@ import 'react-confirm-alert/src/react-confirm-alert.css';
 
 import Axios from '../../blueprints';
 import DataContext from './../../hooks/DataContext';
-import getRandomInt from '../../utils/getRandomInt';
+import getRandomInt from '../../utils/getRandomInt.js';
 
-import { StyledLoadingDots, StyledInput, StyledLink, StyledButton } from './../../styles/';
+import StyledLoadingDots from '../../styles/StyledLoadingDots.jsx';
+import StyledInput from '../../styles/StyledInput.jsx';
+import StyledLink from '../../styles/StyledLink.jsx';
+import StyledButton from '../../styles/StyledButton.jsx';
 import StyledPage from './styles';
 
 function SignIn() {
