+++ src/components/TrendingNav/index.jsx
@@ -1,9 +1,47 @@
 import { useState, useEffect, useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
-
-import { Wrapper, Header, Hashtags, Hashtag } from './styles';
+import styled from 'styled-components';
 import DataContext from '../../hooks/DataContext';
 import Axios from '../../blueprints';
+
+const Wrapper = styled.div`
+  margin-top: 160px;
+  background-color: ${({ theme }) => theme.colors.background};
+  width: 10%;
+  min-width: 300px;
+  border-radius: 16px;
+  * {
+    color: white;
+  }
+`;
+
+const Header = styled.div`
+  font-family: ${({ theme }) => theme.fonts.primary};
+  font-size: 27px;
+  font-weight: 700;
+  line-height: 40px;
+  padding: 12px 16px;
+  border-bottom: ${({ theme }) => theme.styles.defaultBorder};
+`;
+
+const Hashtags = styled.div`
+  padding: 22px 16px;
+  ${({ theme }) => theme.mixins.flexbox('column', 'flex-start', 'flex-start', '5px')};
+`;
+
+const Hashtag = styled.div`
+  font-family: ${({ theme }) => theme.fonts.secondary};
+  font-size: 19px;
+  font-weight: 700;
+  line-height: 23px;
+  letter-spacing: 0.05em;
+  text-align: left;
+  letter-spacing: 0.05em;
+  cursor: pointer;
+  :hover {
+    color: ${({ theme }) => theme.colors.secondary};
+  }
+`;
 
 export default function TrendingNav() {
   const [hashtags, setHashtags] = useState(null);
