--- src/styles/index.jsx
@@ -1,8 +1,0 @@
-import StyledResets from './resets';
-import StyledLoadingDots from './StyledLoadingDots';
-import StyledInput from './StyledInput';
-import StyledLink from './StyledLink';
-import StyledButton from './StyledButton';
-import theme from './theme';
-
-export { StyledResets, StyledLoadingDots, StyledInput, StyledLink, StyledButton, theme };
